Fixed a lot of stuff and deleted a lot of stuff from the scene.
I implemented Robs crystal with animation and replaced the old stand by crystal from the tree tops.

Implemented the particle beam for crystals.

Changed some position of the maple leaf.

**All position of the maple leaf got moved and all the shrine that was done got reset. So i have to implement all of them back. Not sure what happened but now all should be fix and in place.

We are missing some preFab when I was running the game. We don't have the Tilting Log.

Changed a lot of materials and placed a lot of new grasses and few rocks.

Time to sleep.